security things: create directory for reporst if missing

+++ reports_handling.py
@@ -1,6 +1,7 @@
 import datetime
 import database_handling as db
 import csv
+import os
 
 
 def get_datetime_object_from_strings(date_string, time_string):
@@ -20,45 +21,53 @@ def calc_work_time(entry_date, entry_time, exit_date, exit_time):
 
 def create_csv(worker_id):
     if worker_id != '':
-        worker_data = db.get_worker_full_name(db.database_filename, worker_id).split(" ")
-        worker_last_name = worker_data[1]
-        worker_first_name = worker_data[2]
-        filename = ("reports/%s_%s_%s_report.csv" % (worker_id, worker_last_name, worker_first_name))
+        if not os.path.exists("reports"):
+            try:
+                os.mkdir("reports")
+            except OSError:
+                print("Couldn't create directory 'reports'.")
+        if os.path.exists("reports"):
+            worker_data = db.get_worker_full_name(db.database_filename, worker_id).split(" ")
+            worker_last_name = worker_data[1]
+            worker_first_name = worker_data[2]
+            filename = ("reports/%s_%s_%s_report.csv" % (worker_id, worker_last_name, worker_first_name))
 
-        with open(filename, 'w', newline='') as file:
-            writer = csv.writer(file)
+            with open(filename, 'w', newline='') as file:
+                writer = csv.writer(file)
 
-            writer.writerow(
-                ["Entry date", "Entry time", "Entry card", "Entry terminal", "Exit date", "Exit time", "Exit card",
-                 "Exit terminal", "Work time"])
+                writer.writerow(
+                    ["Entry date", "Entry time", "Entry card", "Entry terminal", "Exit date", "Exit time", "Exit card",
+                     "Exit terminal", "Work time"])
 
-            worker_logs = db.get_worker_logs(db.database_filename, worker_id)
+                worker_logs = db.get_worker_logs(db.database_filename, worker_id)
 
-            is_entry = True
-            entry_date = ""
-            entry_time = ""
-            entry_card = ""
-            entry_terminal = ""
+                is_entry = True
+                entry_date = ""
+                entry_time = ""
+                entry_card = ""
+                entry_terminal = ""
 
-            for log in worker_logs:
-                if is_entry:
-                    entry_date = log[0]
-                    entry_time = log[1]
-                    entry_card = log[2]
-                    entry_terminal = log[4]
+                for log in worker_logs:
+                    if is_entry:
+                        entry_date = log[0]
+                        entry_time = log[1]
+                        entry_card = log[2]
+                        entry_terminal = log[4]
 
-                    is_entry = not is_entry
-                else:
-                    exit_date = log[0]
-                    exit_time = log[1]
-                    exit_card = log[2]
-                    exit_terminal = log[4]
-                    work_time = (calc_work_time(entry_date, entry_time, exit_date, exit_time)).__str__()
+                        is_entry = not is_entry
+                    else:
+                        exit_date = log[0]
+                        exit_time = log[1]
+                        exit_card = log[2]
+                        exit_terminal = log[4]
+                        work_time = (calc_work_time(entry_date, entry_time, exit_date, exit_time)).__str__()
 
-                    writer.writerow(
-                        [entry_date, entry_time, entry_card, entry_terminal, exit_date, exit_time, exit_card,
-                         exit_terminal,
-                         work_time])
+                        writer.writerow(
+                            [entry_date, entry_time, entry_card, entry_terminal, exit_date, exit_time, exit_card,
+                             exit_terminal,
+                             work_time])
 
-                    is_entry = not is_entry
-        print("Report %s is ready" % filename)
+                        is_entry = not is_entry
+            print("Report %s is ready" % filename)
+        else:
+            print("Directory 'reports' missing. Please, create directory 'reports' and try again.")
